Add functinality to get logs from a particular perion in time
