--- releng/buildpacks/auditlog-client/src/main/java/com/sap/auditlog/client/config/MissingEnvVariableException.java
@@ -1,8 +1,0 @@
-package com.sap.auditlog.client.config;
-
-public class MissingEnvVariableException extends Exception {
-
-  public MissingEnvVariableException(final String message) {
-    super(message);
-  }
-}
