--- releng/buildpacks/auditlog-client/src/main/java/com/sap/auditlog/client/deserializers/InstantDeserializer.java
@@ -1,15 +1,0 @@
-package com.sap.auditlog.client.deserializers;
-
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-import java.io.IOException;
-import java.time.Instant;
-
-public class InstantDeserializer extends JsonDeserializer<Instant> {
-
-  public Instant deserialize(JsonParser p, DeserializationContext ctxt) throws IOException {
-    return Instant.parse(p.getText());
-  }
-}
-
