--- releng/buildpacks/auditlog-client/src/main/java/com/sap/auditlog/client/exceptions/InvalidMessageException.java
@@ -1,12 +1,0 @@
-package com.sap.auditlog.client.exceptions;
-
-public class InvalidMessageException extends ServiceException {
-
-  public InvalidMessageException(String message) {
-    super(message);
-  }
-
-  public InvalidMessageException(String message, Exception e) {
-    super(message, e);
-  }
-}
