--- releng/buildpacks/auditlog-client/src/main/java/com/sap/auditlog/client/exceptions/ResourceLimitException.java
@@ -1,12 +1,0 @@
-package com.sap.auditlog.client.exceptions;
-
-public class ResourceLimitException extends ServiceException {
-
-  public ResourceLimitException(String message) {
-    super(message);
-  }
-
-  public ResourceLimitException(String message, Exception e) {
-    super(message, e);
-  }
-}
