--- releng/buildpacks/auditlog-client/src/main/java/com/sap/auditlog/client/exceptions/UnauthorizedException.java
@@ -1,12 +1,0 @@
-package com.sap.auditlog.client.exceptions;
-
-public class UnauthorizedException extends ServiceException {
-
-  public UnauthorizedException(String message) {
-    super(message);
-  }
-
-  public UnauthorizedException(String message, Exception e) {
-    super(message, e);
-  }
-}
