--- releng/buildpacks/auditlog-client/src/main/java/com/sap/auditlog/client/http/Communicator.java
@@ -1,14 +1,0 @@
-package com.sap.auditlog.client.http;
-
-import com.sap.auditlog.client.exceptions.ServiceException;
-
-public interface Communicator {
-
-  void send(String apiUrl, String payload, String oauthToken) throws ServiceException;
-
-  String get(String apiUrl, String oauthToken) throws ServiceException;
-
-  String retrieveOAuthToken() throws ServiceException;
-
-  String retrieveOAuthToken(String oauthUrl) throws ServiceException;
-}
