--- releng/buildpacks/auditlog-client/src/main/java/com/sap/auditlog/client/messages/AuditLogCategory.java
@@ -1,8 +1,0 @@
-package com.sap.auditlog.client.messages;
-
-public enum AuditLogCategory {
-  DATA_ACCESS,
-  DATA_MODIFICATION,
-  SECURITY_EVENT,
-  CONFIGURATION_CHANGE
-}
