--- releng/buildpacks/auditlog-client/src/main/java/com/sap/auditlog/client/messages/SecurityAttribute.java
@@ -1,34 +1,0 @@
-package com.sap.auditlog.client.messages;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-@JsonIgnoreProperties(
-    ignoreUnknown = true
-)
-@JsonInclude(Include.NON_NULL)
-@JsonPropertyOrder({"name", "value"})
-public class SecurityAttribute {
-
-  @JsonProperty("name")
-  private final String name;
-  @JsonProperty("value")
-  private final String value;
-
-  public SecurityAttribute(String name, String value) {
-    this.name = name;
-    this.value = value;
-  }
-
-  public String getName() {
-    return this.name;
-  }
-
-  public String getValue() {
-    return this.value;
-  }
-}
-
