+++ releng/buildpacks/auditlog-client/src/main/java/com/sap/xsk/auditlog/client/config/MissingEnvVariableException.java
@@ -1,0 +1,8 @@
+package com.sap.xsk.auditlog.client.config;
+
+public class MissingEnvVariableException extends Exception {
+
+  public MissingEnvVariableException(final String message) {
+    super(message);
+  }
+}
