+++ releng/buildpacks/auditlog-client/src/main/java/com/sap/xsk/auditlog/client/config/ServiceConfig.java
@@ -1,0 +1,32 @@
+package com.sap.xsk.auditlog.client.config;
+
+public class ServiceConfig {
+
+  private final String serviceURL;
+  private final String oauthURL;
+  private final String clientID;
+  private final String clientSecret;
+
+  public ServiceConfig(String serviceURL, String oauthURL, String clientID, String clientSecret) {
+    this.serviceURL = serviceURL;
+    this.oauthURL = oauthURL;
+    this.clientID = clientID;
+    this.clientSecret = clientSecret;
+  }
+
+  public String getServiceURL() {
+    return serviceURL;
+  }
+
+  public String getOauthURL() {
+    return oauthURL;
+  }
+
+  public String getClientID() {
+    return clientID;
+  }
+
+  public String getClientSecret() {
+    return clientSecret;
+  }
+}
