+++ releng/buildpacks/auditlog-client/src/main/java/com/sap/xsk/auditlog/client/deserializers/InstantDeserializer.java
@@ -1,0 +1,16 @@
+package com.sap.xsk.auditlog.client.deserializers;
+
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+import java.io.IOException;
+import java.time.Instant;
+
+public class InstantDeserializer extends JsonDeserializer<Instant> {
+
+  @Override
+  public Instant deserialize(JsonParser p, DeserializationContext ctxt) throws IOException {
+    return Instant.parse(p.getText());
+  }
+}
+
