+++ releng/buildpacks/auditlog-client/src/main/java/com/sap/xsk/auditlog/client/exceptions/InvalidMessageException.java
@@ -1,0 +1,12 @@
+package com.sap.xsk.auditlog.client.exceptions;
+
+public class InvalidMessageException extends ServiceException {
+
+  public InvalidMessageException(String message) {
+    super(message);
+  }
+
+  public InvalidMessageException(String message, Exception e) {
+    super(message, e);
+  }
+}
