+++ releng/buildpacks/auditlog-client/src/main/java/com/sap/xsk/auditlog/client/exceptions/ServiceException.java
@@ -1,0 +1,12 @@
+package com.sap.xsk.auditlog.client.exceptions;
+
+public class ServiceException extends Exception {
+
+  public ServiceException(String message) {
+    super(message);
+  }
+
+  public ServiceException(String message, Exception e) {
+    super(message, e);
+  }
+}
