+++ releng/buildpacks/auditlog-client/src/main/java/com/sap/xsk/auditlog/client/exceptions/UnauthorizedException.java
@@ -1,0 +1,12 @@
+package com.sap.xsk.auditlog.client.exceptions;
+
+public class UnauthorizedException extends ServiceException {
+
+  public UnauthorizedException(String message) {
+    super(message);
+  }
+
+  public UnauthorizedException(String message, Exception e) {
+    super(message, e);
+  }
+}
