+++ releng/buildpacks/auditlog-client/src/main/java/com/sap/xsk/auditlog/client/messages/Attribute.java
@@ -1,0 +1,86 @@
+package com.sap.xsk.auditlog.client.messages;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(Include.NON_NULL)
+public class Attribute {
+
+  @JsonProperty(value = "name", required = true)
+  private final String name;
+  @JsonProperty("successful")
+  private Boolean successful;
+  @JsonProperty("old")
+  private String oldValue;
+  @JsonProperty("new")
+  private String newValue;
+
+  public Attribute(Builder builder) {
+    this.name = builder.name;
+    this.oldValue = builder.oldValue;
+    this.newValue = builder.newValue;
+    this.successful = builder.successful;
+  }
+
+  public String getOldValue() {
+    return this.oldValue;
+  }
+
+  public void setOldValue(String oldValue) {
+    this.oldValue = oldValue;
+  }
+
+  public String getNewValue() {
+    return this.newValue;
+  }
+
+  public void setNewValue(String newValue) {
+    this.newValue = newValue;
+  }
+
+  public String getName() {
+    return this.name;
+  }
+
+  public Boolean getSuccessful() {
+    return this.successful;
+  }
+
+  public void setSuccessful(boolean successful) {
+    this.successful = successful;
+  }
+
+  public static class Builder {
+
+    private final String name;
+    private Boolean successful;
+    private String oldValue;
+    private String newValue;
+
+    public Builder(String name) {
+      this.name = name;
+    }
+
+    public void successful(Boolean successful) {
+      this.successful = successful;
+    }
+
+    public void oldValue(String oldValue) {
+      this.oldValue = oldValue;
+    }
+
+    public void newValue(String newValue) {
+      this.newValue = newValue;
+    }
+
+    public Attribute build() {
+      return new Attribute(this);
+    }
+
+  }
+
+}
+
