+++ releng/buildpacks/auditlog-client/src/main/java/com/sap/xsk/auditlog/client/messages/AuditLogCategory.java
@@ -1,0 +1,8 @@
+package com.sap.xsk.auditlog.client.messages;
+
+public enum AuditLogCategory {
+  DATA_ACCESS,
+  DATA_MODIFICATION,
+  SECURITY_EVENT,
+  CONFIGURATION_CHANGE
+}
