+++ releng/buildpacks/auditlog-client/src/test/java/com/sap/auditlog/client/config/AuditLogReadConfigTest.java
@@ -1,10 +1,10 @@
-package com.sap.auditlog.client.config;
+package com.sap.xsk.auditlog.client.config;
 
 
-import static com.sap.auditlog.client.config.AuditLogReadConfig.OAUTH_API_URL_ENV;
-import static com.sap.auditlog.client.config.AuditLogReadConfig.READ_API_CLIENT_ID_ENV;
-import static com.sap.auditlog.client.config.AuditLogReadConfig.READ_API_CLIENT_SECRET_ENV;
-import static com.sap.auditlog.client.config.AuditLogReadConfig.READ_API_URL_ENV;
+import static com.sap.xsk.auditlog.client.config.AuditLogReadConfig.OAUTH_API_URL_ENV;
+import static com.sap.xsk.auditlog.client.config.AuditLogReadConfig.READ_API_CLIENT_ID_ENV;
+import static com.sap.xsk.auditlog.client.config.AuditLogReadConfig.READ_API_CLIENT_SECRET_ENV;
+import static com.sap.xsk.auditlog.client.config.AuditLogReadConfig.READ_API_URL_ENV;
 
 import org.junit.After;
 import org.junit.Assert;
